+++ CHANGES.rst
@@ -4,4 +4,4 @@ Changes
 0.1.0 (2023-02-05)
 ------------------
 * Initial release.
-* Added support for LCG, FR, .
+* Added support for LCG, FR and two line search function(Armijo & Wolfe).